requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
newspaper3k==0.2.8
nltk==3.8.1
pandas==2.1.1
playwright==1.42.0
tweepy==4.14.0
praw==7.7.1
linkedin-api==2.0.0a5
stackapi==0.2.0
python-dateutil==2.8.2
spacy==3.7.2 